requirements.txt
PyGithub
pandas
giturlparse